Use unit/setUnit/unitName/unitChanged from KoDocument.
Karbon is a bit special since it stores the unit in VDocument too,
but as long as only the part calls VDocument::setUnit, all should be well.

svn path=/trunk/koffice/; revision=472005

+++ karbon/widgets/vcanvas.cc
@@ -279,9 +279,9 @@ VCanvas::setViewportRect( const KoRect &r )
 	viewport()->setUpdatesEnabled( false );
 	double zoomX = m_view->zoom() * ( ( visibleWidth() / m_view->zoom() ) / r.width() );
 	double zoomY = m_view->zoom() * ( ( visibleHeight() / m_view->zoom() ) / r.height() );
-	int pageOffX = ( contentsWidth() - ( m_part->document().width() * m_view->zoom() ) ) / 2.0;
+	double pageOffX = ( contentsWidth() - ( m_part->document().width() * m_view->zoom() ) ) / 2.0;
 	double centerX = double( ( r.center().x() ) * m_view->zoom() + pageOffX ) / double( contentsWidth() );
-	int pageOffY = ( contentsHeight() - ( m_part->document().height() * m_view->zoom() ) ) / 2.0;
+	double pageOffY = ( contentsHeight() - ( m_part->document().height() * m_view->zoom() ) ) / 2.0;
 	double centerY = double( ( r.center().y() ) * m_view->zoom() + pageOffY ) / double( contentsHeight() );
 	double zoom = zoomX < zoomY ? zoomX : zoomY;
 	resizeContents( int( ( zoom / m_view->zoom() ) * contentsWidth() ),
